generator/requirements.txt
shapely==2.0.1
psycopg2-binary==2.9.9
psycopg-binary==3.2.3
tqdm==4.66.1
language_tool_python==2.8
lxml==5.3.0
numpy==2.1.1
bm25s==0.2.1
faiss-cpu==1.8.0
sentence-transformers==3.1.1
pandas
geopandas
langchain==0.3.6
langchain_openai==0.2.4
langchain-ollama==0.2.0
bert-score==0.3.13
geopy==2.4.1